Enable raw inserts in 'dconf set-custom-value' and dict merge as well
With raw inserts the key-value is not stored under "custom_values" but rather inserted straight into the row document. This is good if you know what you are doing.

The key-value is also dict-merged which means it is possible to patch individual values in nested dicts.

The following example modifies the postgres database host for the given tenant:

dconf set-custom-value '{"postgres": {"database": "test-db"}}' --location dg-superkaiju --raw