Make distinction between local and shared config with 'dconf developer' command
The tenant in shared config is called 'developer' and in local config 'localdev'. This prevents name conflicts with resources like  Postgres db's. Before this fix if a local config was recreated it would essentially drop and recreate the Postgres DB used by the shared config as the DB names were identical. This actually happened :/

Also fixing Py3 bug in a print statement.